download widgetjsons, refactor structure
