+++ package.json
@@ -30,6 +30,7 @@
     "date-fns": "^2.30.0",
     "eslint": "^8.53.0",
     "eslint-config-standard": "^17.1.0",
+    "glob": "^10.3.10",
     "node-localstorage": "^3.0.5",
     "winston": "^3.11.0"
   }
