+++ src/api/widget.mjs
@@ -1,0 +1,14 @@
+import { api, jsonContentHeaders } from './api.mjs';
+
+export const getWidgetByID = async (widgetId) => {
+  return await api.get(`/api/widgetType/${widgetId}`);
+};
+
+export const publishWidget = async (widgetJson) => {
+  return await api.post('/api/widgetType', widgetJson, jsonContentHeaders());
+};
+
+// Widget Bundles
+export const getAllWidgetBundles = async () => {
+  return await api.get('/api/widgetsBundles');
+};
